backend/dronecontrol/requirements.txt
channels[daphne]==4.2.0
channels_redis==4.2.1
django==5.1.4
django-environ==0.11.2
djangorestframework==3.15.2
django-cors-headers==4.7.0
psycopg2==2.9.10
Shapely==2.0.7
geopandas==1.0.1
numpy==1.23.5
pytest==8.3.4
pytest-asyncio==0.25.3
pytest-django==4.10.0
Rtree==1.3.0
